requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==13.1.0
beautifulsoup4==4.13.4
black==25.1.0
click==8.1.8
cloup==3.0.7
colorama==0.4.6
decorator==5.2.1
exceptiongroup==1.3.0
fastapi==0.115.12
flake8==7.2.0
glcontext==3.0.0
h11==0.16.0
idna==3.10
isort==6.0.1
isosurfaces==0.1.2
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
moderngl==5.12.0
moderngl-window==3.1.1
mypy_extensions==1.1.0
networkx>=3.0
numpy>=1.24.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pycairo>=1.20.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.3.2
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
rich==14.0.0
scipy>=1.10.0
screeninfo==0.8.1
skia-pathops==0.8.0.post2
sniffio==1.3.1
soupsieve==2.7
srt==3.5.3
starlette==0.46.2
svgelements==1.9.6
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchdog==6.0.0
requests==2.31.0
python-dotenv==1.0.1
supabase==2.9.1
pyjwt==2.9.0
email-validator==2.1.0

importlib-metadata==8.5.0
